scripts/requirements_enterprise.txt
# Enterprise Document Extraction API Requirements
# Compatible with Python 3.8+

# Core FastAPI stack
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0.post1
gunicorn==21.2.0

# Authentication and rate limiting
slowapi==0.1.9
python-multipart==0.0.6

# Caching and data
redis==5.0.1
hiredis==2.2.3

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Resilience and reliability
tenacity==8.2.3
httpx==0.25.2

# File processing
aiofiles==23.2.1
python-magic==0.4.27

# Data validation and serialization
pydantic[email]==2.4.2
pydantic-settings==2.0.3

# Async and concurrency
asyncio-throttle==1.0.2

# PDF processing (from existing requirements)
PyPDF2==3.0.1
pdfplumber==0.10.3

# Office documents (from existing requirements)
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.23

# Web documents (from existing requirements)
beautifulsoup4==4.12.2
lxml==4.9.3

# Text encoding detection (from existing requirements)
chardet==5.2.0

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Security
cryptography==41.0.7
psutil>=5.9.0
psutil>=5.9.0
